move profile command into own package
