+++ src/main/java/com/fossgalaxy/pircbotx/commands/profile/ProfileModel.java
@@ -16,9 +16,8 @@
  * You should have received a copy of the GNU General Public License along with
  * uc_pircbotx. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.fossgalaxy.pircbotx.data.db;
+package com.fossgalaxy.pircbotx.commands.profile;
 
-import com.fossgalaxy.pircbotx.profile.Profile;
 import com.google.inject.Inject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
